Add shared helpers for parsing line protocol
- tags.py, fields.py, measurements.py, and tag_to_field.py now import shared helpers instead of duplicating scanners. The repeated temp-file rewrite blocks were consolidated into _rewrite_file_in_place, which makes the mutating commands much more uniform. The command-level logic stays local, but the parsing primitives now live in one place, which should make future fixes and tuning safer.
